ai_correction/requirements.txt
streamlit>=1.28.0
PyMuPDF>=1.25.0
Pillow>=10.0.0
fastapi>=0.104.1
uvicorn>=0.24.0
python-multipart>=0.0.6
pydantic>=2.5.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
reportlab>=4.0.8
weasyprint>=62.0
jinja2>=3.1.2
markdown>=3.5.1
python-docx>=1.1.0
openpyxl>=3.1.2
requests>=2.31.0
openai>=1.45.0
PyJWT>=2.8.0

# LangGraph 依赖
langgraph>=0.0.20
langchain>=0.1.0
langchain-core>=0.1.0

# 数据库依赖
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0  # PostgreSQL
pymysql>=1.1.0  # MySQL
alembic>=1.13.0  # 数据库迁移

# Google Gemini 3 Pro SDK (官方 SDK)
google-genai>=0.2.0

# 图像处理依赖（已移除，不再需要）
# numpy>=1.24.0
# opencv-python>=4.8.0